fix: 收窄终结失败缓存范围
